+++ releaseNotes.md
@@ -1,3 +1,13 @@
+_October 20th, 2019_  **v2.1.22**
+- Pagination - Seems when query params OTHER than pagination query parms are sent, then one the results were less than max alloted, the **NEXT** was being populated and should not have been. 
+
+Thanks,  
+_Slacquer_  
+
+<br>
+<br>
+
+
 _August 17th, 2019_  **v2.1.21**
 - APIBlox.AspNetCore.CommandsQueriesControllersOhMy - Templates, up until now, all handlers were injected into the controllers constructor.  Not a big deal unless the controller gets quite large, therefore all handlers are now injected into the actions themselves ([FromServices] attribute).  This should help prevent the need to seperate the controllers into their own files.